refactor(utils): 将网络状态监听功能移动到utils目录
将listenNetworkStatus函数从原位置移动到utils/network.js，并更新相关引用
同时添加utils目录到vite自动导入配置中
