+++ vite.config.js
@@ -25,7 +25,7 @@ export default defineConfig(({ mode }) => {
         dts: 'src/auto-imports.d.ts',
         dirs: [
           'src/composables/**',
-          // 'src/store/**',
+          'src/utils/**',
         ],
         vueTemplate: true,
       }),
